Clean up test files for schema
